cdc_consumer/pyproject.toml
[tool.poetry]
name = "cdc-consumer"
version = "0.1.0"
description = ""
authors = ["rahul@apperture.io"]
readme = "README.md"
packages = [{include = "cdc_consumer"}]

[tool.poetry.dependencies]
python = "^3.9"
aiokafka = "^0.8.0"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
gunicorn = "^20.1.0"
clickhouse-connect = "0.3.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
flatten-json = "^0.1.13"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
